feat: 교환 신청 시 기존 요청이 있는 경우 TradeRequest에 데이터 전달
